Rename notebook.py to Project: Do Left-handed People Really Die Young?
In this project, I explored this phenomenon using age distribution data to see if I can reproduce a difference in average age at death purely from the changing rates of left-handedness over time, refuting the claim of early death for left-handers. In this notebook I have used pandas and Bayesian statistics to analyse the probability of being a certain age at death given that you are reported as left-handed or right-handed.